[suiop] add point of contact selection for incident review
