+++ crates/suiop-cli/Cargo.toml
@@ -51,6 +51,7 @@ toml_edit.workspace = true
 tracing-subscriber.workspace = true
 tracing.workspace = true
 once_cell.workspace = true
+futures.workspace = true
 strsim = "0.11.1"
 
 [dev-dependencies]
