+++ crates/suiop-cli/src/cli/incidents/mod.rs
@@ -4,6 +4,7 @@
 mod jira;
 mod pd;
 mod slack;
+mod slack_api;
 
 use anyhow::Result;
 use chrono::{Duration, Local};
